% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctions.R
\name{circleFun}
\alias{circleFun}
\title{Function to plot circles in ggplot}
\usage{
circleFun(
  rx = 1,
  ry = 1,
  start = 0,
  end = 360,
  length = 360,
  xstart = 0,
  ystart = 0
)
}
\arguments{
\item{rx}{Radius of the circle in x axis}

\item{ry}{Radius of the circle in y axis}

\item{start}{Angle (in degrees) from where to start drawing (from vertical)}

\item{end}{Angle (in degrees) to end the drawing}

\item{xstart}{X coordinate of the origin of the circle}

\item{ystart}{Y coordinate of the origin of the circle}
}
\description{
Function to plot circles in ggplot
}
