% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTargets.R
\name{createTargets}
\alias{createTargets}
\title{Create target file}
\usage{
createTargets(
  names,
  column_list = NULL,
  fileName = "targets.txt",
  resultsDir = getwd()
)
}
\arguments{
\item{names}{Vector of the sample names to be analyzed. Usually the column names of the count table}

\item{column_list}{Named list of the columns to be added to the targets file, defining the sample. Usually treatment, timepoint, condition... Default = NULL}

\item{fileName}{Name of the output file, with the extension. Default = "targets.txt"}

\item{resultsDir}{Output directory. Default = current working directory}
}
\value{
Saves a tabulated table at resultsDir, with the columns of the conditions are provided as a named list, or an empty file if nothing is provided (to be filled manually later).
}
\description{
Creates an empty targets file with the specified names. If a named list of variables is provided, creates the filled targets file.
}
\examples{
# prepare columns
treat <- c("Untreat","Untreat","Untreat","Treat","Treat","Treat") 
weeks <- c("4w","9w","13w","4w","9w","13w")
cond <- paste(treat,weeks, sep=".")
# create list of the columns, must be with names for proper conversion to column name
column_list <- list(Treat=treat,Weeks=weeks,Cond=cond)

createTargets(names=colnames(countsM), column_list = column_list, resultsDir=resultsDir)

# after the file has created, we read it as table
targets <- read.table(file.path(resultsDir,"targets.txt"), dec=",", sep="\t", header=T)
}
