% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctions.R
\name{get_lower_tri}
\alias{get_lower_tri}
\title{Get lower triangle of a square matrix (for example the correlation matrix) to plot with ggplot.}
\usage{
get_lower_tri(mat)
}
\arguments{
\item{mat}{Squared Matrix}
}
\description{
Get lower triangle of a square matrix (for example the correlation matrix) to plot with ggplot.
}
\examples{
cor(longley) # Original Matrix
get_lower_tri(cor(longley)) # Lower triangle
}
