% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAnnot.R
\name{getAnnot}
\alias{getAnnot}
\title{Gets annotation dataframe of a filtered counts matrix for an already created annotation file, that will be dependent on the genome species and version}
\usage{
getAnnot(counts, genome, annotDir)
}
\arguments{
\item{counts:}{Filtered count matrix}

\item{genome:}{Genome that was used to create the count matrix. Example: mm10, GRCh38....}

\item{annotDir:}{Directory where the annotation files are stored}
}
\value{
Annotation dataframe of the genes included in the counts matrix
}
\description{
Gets annotation dataframe of a filtered counts matrix for an already created annotation file, that will be dependent on the genome species and version
}
