% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctions.R
\name{ggplotly_slides}
\alias{ggplotly_slides}
\title{Interactive plot example to use when plotly legend is not the same as in ggplot2}
\usage{
ggplotly_slides(x, position = "bottom", col_width = 10)
}
\arguments{
\item{x}{Plot generated with ggplot2}

\item{position}{Position of the legend (only right is implemented)}

\item{col_width}{Width of the first column. Second column width is 12-col_width}
}
\description{
Interactive plot example to use when plotly legend is not the same as in ggplot2
}
