% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctions.R
\name{\%!idx\%}
\alias{\%!idx\%}
\title{Function to find element INDICES in 'x' different than 'y'}
\usage{
x \%!idx\% y
}
\arguments{
\item{x}{Vector to look for}

\item{y}{Elements to exclude from 'x'}
}
\description{
Function to find element INDICES in 'x' different than 'y'
}
\examples{
x<-c(1,2,3);y<-c(3)
x \%!in\% y
}
\keyword{!in}
