% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctions.R
\name{\%!in\%}
\alias{\%!in\%}
\title{Function to find elements (TRUE or FALSE) in 'x' different than 'y'. To find indices, use %!idx% instead}
\usage{
x \%!in\% y
}
\arguments{
\item{x}{Vector to look for}

\item{y}{Elements to exclude from 'x'}
}
\description{
Function to find elements (TRUE or FALSE) in 'x' different than 'y'. To find indices, use %!idx% instead
}
\examples{
x<-c(1,2,3);y<-c(3)
x \%!in\% y
}
\keyword{!in}
