% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCls.R
\name{makeCls}
\alias{makeCls}
\title{Creates a cls file using a vector of conditions, to be used in GSEA}
\usage{
makeCls(cond, resultsDir = getwd(), fileName)
}
\arguments{
\item{cond:}{column of the conditions to be compared of the targets file}

\item{resultsDir:}{output directory. Default = working directory}

\item{fileName:}{name of the output file, without extension}
}
\value{
cls file with the specific structure in specified path
}
\description{
Creates a cls file using a vector of conditions, to be used in GSEA
}
