% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGSEA.R
\name{makeGSEA}
\alias{makeGSEA}
\title{Runs Gene set enrichment analysis (GSEA) using package clusterProfiler}
\usage{
makeGSEA(
  results,
  contrast,
  gmt,
  resultsDir = getwd(),
  collectionName = "",
  minGSSize = 15,
  maxGSSize = 500,
  p.value = 1,
  p.adj = 0.05,
  rankMethod = "sign",
  plots = TRUE,
  plots2Make = c("Barplot", "Dotplot", "RunScoreSig", "GCNetwork", "EnrichMAP"),
  plotTop = 30,
  plotTopRS = 50,
  JoinedDotplot = TRUE,
  contrast.order = NULL,
  leadingEdges = FALSE,
  ...
)
}
\arguments{
\item{results}{ResultsRNAseq object, a data.frame obtained from `RNAseq.resAnnot()`. 
Must contain a Geneid column and ##' P.Value, adj.P.Val and logFC columns for each contrast. Example: "P.Value.LES_ACT.vs.LES_INACT"}

\item{contrast}{List with one vector of length 2 for each contrast.}

\item{gmt}{Gene sets data.frame as obtained with function `clusterProfiler::read.gmt`. Must have a column "term" and a column "gene".}

\item{resultsDir}{Character vector with output results directory. Default = working directory}

\item{collectionName}{Vector with the name of the collection. It will be 
appended to the output directory/files name. Example: "c5.go.bp". Default = ""}

\item{minGSSize}{Minimal size of each geneSet for analyzing. Default = 15}

\item{maxGSSize}{Maximal size of genes annotated for testing. Default = 500}

\item{p.value}{p-value cutoff for RNAseq.resAnnot objects. Default = 1, as we want to keep all results}

\item{p.adj}{Threshold of the adjusted p-value to be selected for plots. Default = 0.05}

\item{rankMethod}{By default, the ranking is performed with the log p-value and the sign of fold change ("sign"). However, ranking by both log p-value and logFC is possible when specifying "logFC". Default = "sign"}

\item{plots}{Whether to generate plots (bar plot, dot plot, enrichment map and gene concept networks) in the same directory. Default = TRUE}

\item{plots2Make}{List of plots that will be performed by makeGSEAplots, if plots = TRUE. By default, all of them will be performed. Default =  c("Barplot", "Dotplot", "RunScoreSig", "GCNetwork", "EnrichMAP")}

\item{plotTop}{Number of maximum gene sets to be plotted in barplot and dotplot, and enrichment map. Default = 30}

\item{plotTopRS}{Number of maximum plots to perform the running score. Default = 50}

\item{JoinedDotplot}{Whether to perform a joined dotplot that compares the overlap of the enriched terms between comparisons. Default = TRUE}

\item{contrast.order}{Order in which the contrasts will be shown in the JoinedDotplot. If NULL, it will use the order as shown in contrast. Default = NULL}

\item{leadingEdges}{Whether to make an excel including how many and which gene sets a significant gene is part of. Will make a folder "leadingEdges" with the excels per comparison. Default = FALSE}

\item{...}{Additional parameters}
}
\value{
Returns a list with enrichment results for each contrast. The list is saved as a RData object 
in the resultsDir directory, along with a folder for each contrast containing an 
excel file with enrichment results and plots if requested.
}
\description{
Function that performs a GSEA on a RNAseq.resAnnot object (and a list of contrasts) or a genelist
from clusterProfiler package. This function performs an hypergeometric test for each gene set.
}
