% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLeadingEdges.R
\name{makeLeadingEdges}
\alias{makeLeadingEdges}
\title{Runs Leading Edges analysis from Gene set enrichment analysis (GSEA)}
\usage{
makeLeadingEdges(
  gsea,
  contrast,
  p.adj = 0.05,
  collectionName = "",
  gseaDir,
  ...
)
}
\arguments{
\item{gsea}{list of GSEA objects (of length n contrasts)}

\item{contrast}{List with one vector of length 2 for each contrast.}

\item{p.adj}{Adjusted p-value threshold of significance. Default = 0.05}

\item{collectionName}{Vector with the name of the collection. It will be appended to the output directory/files name. Example: "H". Default = ""}

\item{gseaDir}{Directory to save the results}

\item{...}{Additional parameters}
}
\value{
Excel with the genesets in which a significant gene is part of
}
\description{
Runs Leading Edges analysis from Gene set enrichment analysis (GSEA)
}
