% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeORA.R
\name{makeORA}
\alias{makeORA}
\title{Runs over representation analysis using enrichr function of cluterProfiler}
\usage{
makeORA(
  results,
  contrast = NULL,
  gmt,
  collectionName = "",
  resultsDir = getwd(),
  minGSSize = 15,
  maxGSSize = 500,
  p.value = 0.05,
  p.adjust = 0.05,
  logFC = 0,
  plots = TRUE,
  plotTop = 50,
  plotP.adjust = 0.05
)
}
\arguments{
\item{results}{resultsRNAseq object or a genelist (with names). resultsRNAseq is a data.frame obtained from resultsRNAseq(). 
Must contain a Geneid column and ##' P.Value, adj.P.Val and logFC columns for each contrast (eg. "P.Value.LES_ACT.vs.LES_INACT"). 
ORA will be performed on the genelist if contrast parameter is NULL}

\item{contrast}{List with one vector of length 2 for each contrast. Default = NULL}

\item{gmt}{Gene sets data.frame as obtained with function clusterProfiler::read.gmt. Must have a column "term" and a column "gene".}

\item{collectionName}{Vector with the name of the collection. It will be 
appended to the output directory/files name (eg. "c5.go.bp")}

\item{resultsDir}{Character vector with output results directory. Default is working directory.}

\item{minGSSize}{minimal size of each geneSet for analyzing. Default = 15}

\item{maxGSSize}{maximal size of genes annotated for testing. Default = 500}

\item{p.value}{pvalue cutoff for resultsRNASeq objects. Default = 0.05}

\item{p.adjust}{adjusted pvalue cutoff for resultsRNASeq objects. Default = 0.05}

\item{logFC}{logFC cutoff for resultsRNASeq objects. Default = 0}

\item{plots}{whether to generate plots (bar plot, dot plot, enrichment map and gene concept networks) in the same directory. Default = TRUE}

\item{plotTop}{Number of maximal gene sets to be plotted in barplot and dotplot. Default = 50}

\item{plotP.adjust}{threshold of the adjusted p-value to be selected for plots. Default = 0.05}
}
\value{
Returns a list with enrichment results for each contrast. The list is saved as a RData object 
in the resultsDir directory, along with a folder for each contrast containing an 
excel file with enrichment results.
}
\description{
Function that performs a gene-set enrichment analysis using the function enricher
on a resultsRNAseq object (and a list of contrasts) or a genelist
from clusterProfiler package. This function performs an hypergeometric test for each gene set.
}
