% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGSEA.R
\name{makePlotsGSEA}
\alias{makePlotsGSEA}
\title{Generates plots of Gene set enrichment analysis (GSEA)}
\usage{
makePlotsGSEA(
  gsea,
  contrast,
  collectionName = "",
  resultsDir = getwd(),
  Barplot = TRUE,
  Dotplot = TRUE,
  RunScoreSig = TRUE,
  GCNetwork = TRUE,
  EnrichMAP = TRUE,
  GeneSetRS = NULL,
  JoinedDotplot = TRUE,
  plotTop = 50,
  p.adj = 0.05,
  contrast.order = contrast.t,
  ...
)
}
\arguments{
\item{gsea}{GSEA object generated with clusterProfiler. Returned from the `makeGSEA` function.}

\item{contrast}{List with one vector of length 2 for each contrast.}

\item{collectionName}{Vector with the name of the collection.}

\item{resultsDir}{Character vector with output results directory. Default = working directory}

\item{Barplot}{Whether to perform the general barplot of significant results. If there are more than 30 significant gene sets up or down, only top 30 will be added. Default = TRUE}

\item{Dotplot}{Whether to perform GSEA dotplot. First 40 terms will be displayed. Default = TRUE}

\item{RunScoreSig}{Whether to perform the running scores of all significant terms. A maximum of 50 plots are generated by default. This number can be customized with parameter `plotTop`. Default = TRUE}

\item{GCNetwork}{Whether to perform gene-concept network. Displays top 5 gene sets by default. Default = TRUE}

\item{EnrichMAP}{Whether to perform enrichment map. It uses Jaccard index, and plots top 30 by default. Default = TRUE}

\item{GeneSetRS}{Name of gene set or list of gene sets to perform the running score instead of all significant. Default = NULL}

\item{JoinedDotplot}{Whether to perform the joined dotplot of common terms between all contrasts. Can also be obtained in separate function `makeJoinedDotplot`. Default = TRUE}

\item{plotTop}{Number of maximum plots to perform the running score. Default = 50}

\item{p.adj}{Threshold of the adjusted p-value to be selected for plots. Default = 0.05}

\item{contrast.order}{Vector with the order in which the contrasts will be shown in JoinedDotplot}

\item{...}{Addition parameters}
}
\value{
Returns plots for each comparison in each correspondent subfolder at resultsDir
}
\description{
Function that generates the plots to interpret GSEA results, some of them custom and some from the clusterProfiler package. All the following plots are generated by default: a general barplot, a dotplot, running scores, gene-concept networks, enrichment map, and a joined dotplot of common genesets in all conditions.
}
