% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctions.R
\name{normalize}
\alias{normalize}
\title{Normalize based on min and max value.}
\usage{
normalize(x, scale = 1)
}
\arguments{
\item{x}{Numeric vector}

\item{scale}{Scaling factor}
}
\description{
Normalize based on min and max value.
}
\examples{
# x<-rnorm(10,10,1)
# normalize(x)
}
