% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normTMM.R
\name{normTMM}
\alias{normTMM}
\title{Function to normalize to TMMs a counts table using edgeR package
To be used for visualization, not DEA}
\usage{
normTMM(countsTable, log = FALSE, ...)
}
\arguments{
\item{countsTable}{table of counts, with genes as rows and samples as columns}

\item{log}{to be used in function cpm. Default = FALSE}
}
\value{
Normalized table of counts
}
\description{
Function to normalize to TMMs a counts table using edgeR package
To be used for visualization, not DEA
}
