% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctions.R
\name{not_idx}
\alias{not_idx}
\alias{\%!idx\%}
\title{Function to find elements of 'x' elements different than 'y'.}
\usage{
x \%!idx\% y
}
\arguments{
\item{x}{Vector to look for}

\item{y}{Elements to exclude from 'x'}
}
\description{
Function to find elements of 'x' elements different than 'y'.
}
\examples{
x<-c(1,2,3);y<-c(3)
x \%!idx\% y
}
\keyword{!in}
