% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeORA.R
\name{plotORA}
\alias{plotORA}
\title{plot ORA results, internal function}
\usage{
plotORA(
  enrichment,
  plotName,
  collectionName = "",
  resultsDir = getwd(),
  plotTop = 50,
  plotP.adjust = 0.05
)
}
\arguments{
\item{enrichment}{results of an ORA enrichment performed with function makeORA}

\item{plotName}{the name to be given to that plot}

\item{collectionName}{name of the mSigDB collection.}

\item{resultsDir}{Character vector with output results directory}

\item{plotTop}{Number of maximal gene sets to be plotted in barplot and dotplot. Defalult = 50}

\item{plotP.adjust}{threshold of the adjusted p-value to be selected for plots. Default = 0.05}
}
\description{
plot ORA results, internal function
}
