% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctions.R
\name{set_sourceFile}
\alias{set_sourceFile}
\title{Set working Directory to current file folder}
\usage{
set_sourceFile(name = NULL)
}
\description{
Set working Directory to current file folder
}
\examples{
# set_sourceFile
}
