% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctions.R
\name{source_wd}
\alias{source_wd}
\title{Set source file directory as working directory}
\usage{
source_wd(name = NULL)
}
\arguments{
\item{name}{R script file name}
}
\description{
Set source file directory as working directory
}
\examples{
source_wd()
}
