% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctions.R
\name{updatePackage}
\alias{updatePackage}
\title{Function to update a package based on its name.}
\usage{
updatePackage(pckgeName = NULL)
}
\arguments{
\item{pckgeName}{Name of the package to update. Your working directory should be the parent folder of the package}
}
\description{
Function to update a package based on its name.
}
\examples{
updatePackage()
}
