% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctions.R
\name{upper_tri}
\alias{upper_tri}
\title{Get upeer triangle of a square matrix (for example the correlation matrix) to plot with ggplot.}
\usage{
upper_tri(mat)
}
\arguments{
\item{mat}{Squared Matrix}
}
\description{
Get upeer triangle of a square matrix (for example the correlation matrix) to plot with ggplot.
}
\examples{
cor(longley) # Original Matrix
upper_tri(cor(longley)) # Upper triangle
}
