% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTargets.R
\name{createTargets}
\alias{createTargets}
\title{Creates an empty targets file with the specified names. If a named list of variables is provided, creates the filled targets file.}
\usage{
createTargets(
  names,
  column_list = NULL,
  fileName = "targets.txt",
  resultsDir = getwd()
)
}
\arguments{
\item{names:}{usually the sample names to be analyzed}

\item{column_list:}{named list of the columns to be added to the targets file, defining the sample. Usally treatment, timepoint, condition... Default = NULL}

\item{resultsDir:}{Output directory. Default = current working directory}

\item{fileName:}{name of the output file, with extension. Default = "targets.txt"}
}
\value{
empty targets file in resultsDir directory, or full targets file if named list of the columns is provided
}
\description{
Creates an empty targets file with the specified names. If a named list of variables is provided, creates the filled targets file.
}
