% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeRNAseqResults.R
\name{makeExcelResults}
\alias{makeExcelResults}
\title{Create RNAseq results in Excel}
\usage{
makeExcelResults(
  pathRDS,
  fileRDS,
  contrast,
  resultsDir,
  fileName,
  pvalue = NULL,
  padj = 0.05,
  logFC = 1,
  add.colors = NULL
)
}
\arguments{
\item{pathRDS}{path to RDS file with a data.frame obtained from `RNAseq.resAnnot()` object}

\item{fileRDS}{RDS file with a data.frame obtained from resultsRNAseq() object, without extension}

\item{contrast}{List of vectors with each contrast to use}

\item{resultsDir}{Output directory}

\item{fileName}{Name of the output file, without extension}

\item{pvalue}{p-value threshold. Default = NULL, as it is assumed to use an adjusted p-value by default}

\item{padj}{adjusted p-value threshold. Default = 0.05}

\item{logFC}{abs(logFC) threshold. Default = 1}

\item{add.colors}{colors to add if there are more than 20 contrasts. Default = NULL}
}
\value{
Excel file in resultsDir with fileName
}
\description{
Creates an excel file out of a ResultsRNAseq file
}
\seealso{
[RNAseq.resAnnot()]
}
