% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGSEA.R
\name{makeGSEA}
\alias{makeGSEA}
\title{Runs Gene set enrichment analysis (GSEA) using package clusterProfiler}
\usage{
makeGSEA(
  results,
  contrast,
  gmt,
  resultsDir = getwd(),
  collectionName = "",
  minGSSize = 15,
  maxGSSize = 500,
  p.value = 1,
  plots = TRUE,
  p.adj = 0.05,
  plotTop = 50,
  leadingEdges = FALSE,
  ...
)
}
\arguments{
\item{results}{ResultsRNAseq object or a genelist (with names). resultsRNAseq is a data.frame obtained from `RNAseq.resAnnot()`. 
Must contain a Geneid column and ##' P.Value, adj.P.Val and logFC columns for each contrast. Example: "P.Value.LES_ACT.vs.LES_INACT"}

\item{contrast}{List with one vector of length 2 for each contrast.}

\item{gmt}{Gene sets data.frame as obtained with function `clusterProfiler::read.gmt`. Must have a column "term" and a column "gene".}

\item{resultsDir}{Character vector with output results directory. Default = working directory}

\item{collectionName}{Vector with the name of the collection. It will be 
appended to the output directory/files name. Example: "c5.go.bp"}

\item{minGSSize}{Minimal size of each geneSet for analyzing. Default = 15}

\item{maxGSSize}{Maximal size of genes annotated for testing. Default = 500}

\item{p.value}{p-value cutoff for RNAseq.resAnnot objects. Default = 1, as we want to keep all results}

\item{plots}{Whether to generate plots (bar plot, dot plot, enrichment map and gene concept networks) in the same directory. Default = TRUE}

\item{p.adj}{Threshold of the adjusted p-value to be selected for plots. Default = 0.05}

\item{plotTop}{Number of maximal gene sets to be plotted in barplot and dotplot. Default = 50}

\item{leadingEdges}{Whether to make an excel including how many and which gene sets a significant gene is part of. Will make a folder "leadingEdges" with the excels per comparison. Default = FALSE}

\item{...}{Additional parameters}
}
\value{
Returns a list with enrichment results for each contrast. The list is saved as a RData object 
in the resultsDir directory, along with a folder for each contrast containing an 
excel file with enrichment results and plots if requested.
}
\description{
Function that performs a GSEA on a RNAseq.resAnnot object (and a list of contrasts) or a genelist
from clusterProfiler package. This function performs an hypergeometric test for each gene set.
}
