% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGSEA.R
\name{makeJoinedDotplot}
\alias{makeJoinedDotplot}
\title{Make GSEA dotplot}
\usage{
makeJoinedDotplot(
  gsea,
  contrast,
  p.adj = 0.05,
  collectionName = "",
  resultsDir = getwd()
)
}
\arguments{
\item{gsea}{GSEA results object.}

\item{contrast}{List of vectors with each contrast to use}

\item{p.adj}{Adjusted p-value threshold. Default = 0.05}

\item{collectionName}{Name of the collection used for the GSEA, e.g. "H" for Hallmark. Default = ""}

\item{resultsDir}{Output directory where results will be stored. Default = current working directory}
}
\value{
Saves a dotplot of GSEA significant results in all conditions, as well as an excel with the same information
}
\description{
Generates a dotplot of GSEA results joining all comparisons when there are less than 100 significant genesets results, and an equivalent heatmap in Excel format for all cases
}
\seealso{
[makeGSEA()]
}
